Added queued action functionality to DHT querying.
