+++ src/main/kotlin/network/DistributedHashTable.kt
@@ -10,6 +10,8 @@ import data.network.Node
 import kotlinx.serialization.encodeToByteArray
 import kotlinx.serialization.protobuf.ProtoBuf
 import logging.Logger
+import utils.launchCoroutine
+import java.util.concurrent.ConcurrentHashMap
 
 /**
  * Created by Mihael Valentin Berčič
@@ -18,11 +20,20 @@ import logging.Logger
  */
 abstract class DistributedHashTable(configuration: Configuration) : Server(configuration) {
 
-    fun queryFor(vararg publicKeys: String) {
+    private val queuedActions = ConcurrentHashMap<String, (Node) -> Unit>()
+
+    fun queryFor(vararg publicKeys: String, onFoundBlock: ((Node) -> Unit)? = null) {
         val unknown = publicKeys.filter { !knownNodes.containsKey(it) }
         if (unknown.isNotEmpty()) {
             val queryMessage = QueryMessage(localNode, publicKeys)
             send(Endpoint.NodeQuery, TransmissionType.Unicast, queryMessage)
+        }
+        if (onFoundBlock != null && publicKeys.isNotEmpty()) {
+            val known = publicKeys.mapNotNull { knownNodes[it] }
+            queuedActions.putAll(unknown.map { it to onFoundBlock })
+            known.forEach { node ->
+                launchCoroutine { onFoundBlock(node) }
+            }
         }
     }
 
@@ -50,11 +61,18 @@ abstract class DistributedHashTable(configuration: Configuration) : Server(confi
         val foundNodes = message.decodeAs<Array<Node>>()
         addNewNodes(*foundNodes)
         checkForQueuedMessages(foundNodes)
+        invokeQueuedAction(*foundNodes)
     }
 
     fun addNewNodes(vararg nodes: Node) {
         val mapped = nodes.associateBy { it.publicKey }
         knownNodes.putAll(mapped)
+    }
+
+    private fun invokeQueuedAction(vararg nodes: Node) {
+        nodes.forEach { node ->
+            queuedActions.remove(node.publicKey)?.invoke(node)
+        }
     }
 
     inline fun <reified T> send(endpoint: Endpoint, transmissionType: TransmissionType, data: T, vararg publicKeys: String) {
